Разделяет код по модулям
